CustomQueryView merged
